Local Setup
NO CHANGES to the original ,code from Mike Primm.
I just Forked this repro and plan of changing it for my needs.
