--- src/main/resources/config.yml
@@ -1,5 +1,0 @@
-# SchematicBrush configuration
-schematic-sets:
-
-
-
